Fix step logic in engine and move switchSensor in sensor.go
